requirements.txt
argcomplete==3.4.0
cmake==3.30.5
contourpy==1.3.0
cycler==0.12.1
filelock==3.16.1
fonttools==4.54.1
importlib_resources==6.4.5
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.7
lit==18.1.8
MarkupSafe==3.0.2
matplotlib==3.9.2
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
packaging==24.1
pandas==2.2.3
pillow==11.0.0
pipx==1.6.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
scikit-learn==1.4.2
scipy==1.13.1
seaborn==0.13.2
six==1.16.0
sympy==1.13.3
threadpoolctl==3.5.0
tomli==2.0.1
torch==2.0.1
triton==2.0.0
typing_extensions==4.12.2
tzdata==2024.2
userpath==1.9.2
zipp==3.20.2
